Simplify `JsonSchema#inspect`
Unfortunately, our recursive inspect seems to cause trouble in some
cases (see interagent/committee#77). Even when it doesn't, it's quite
expensive to produce and can cause some trouble when higher level
libraries call it automatically.

This simplifies `#inspect` to produce a minimal representation. The
helper in `#inspect_schema` is still available for debugging.