Create a mechanism for normalized release with Bundler's tasks
I always forget to push commits and/or tags and/or gems during a release, so let's normalize the process using Bundler's built-in helpers for it.

One caveat is that the format of the tag changes from "1.10.0" to one with a prefixed "v" like "v1.10.0", but that shouldn't matter all that much in the long run.

Almost a perfect copy of a sister pull in interagent/committee#89.