Expand a schema that is _just_ a reference
The current reference expander does a pretty good job of expanding
references that exist as subschemas in the schema, but fails to expand a
schema that is _just_ a JSON reference and nothing else.

This adds some code to short circuit the expansion mechanism if we
detect a simple JSON reference and nothing else.