Assign links a fragment so they can be inspected
Previously, links were not being assigned a fragment (but were assigned
a parent) so if `#inspect` (which calls `#pointer`) was invoked on them,
we'd throw an exception as we tried to concatenate a string to `nil`.

This patch addresses that issue by having the parser assign links a
fragment so that a pointer can successfully be built for them.

Fixes #82.