Add schema_name for max/min + fix others
Adds a schema name for the `max` and `min` properties of a schema so
that when we index them with `#[]`, we'll get back the value that we
expect (for example, maximum should be accessible with `[:maximum]`.

Also fix the schema names for `max_exclusive` and `min_exclusive` which
were entered incorrectly.