Reduce hash insertion dups
The string return value of `pointer` is used as a hash key:

  https://github.com/brandur/json_schema/blob/bce0056f47a5d4114668702352d185a5b05dc61f/lib/json_schema/reference_expander.rb#L72

Hashes will dup key strings in case the original string gets mutated.
If we freeze the return value of `pointer`, the hash knows the string
can't be mutated, so the string value is used directly as the key.

This saves about 1,160,383 allocations.

Before:

```
{:REAL=>4.689677000045776}
{:ALLOCATIONS=>6167045}
```

After:

```
{:REAL=>4.527120999991894}
{:ALLOCATIONS=>5006662}
```