Remove Ruby 2.0 and 2.1 from build matrix
It seems that the Bundler project has started to aggressively drop
support for older Ruby versions, and 2.0 and 2.1 are among them. It's
difficult to continue supporting them when Bundler has dropped them, so
here we drop them too.

Both have been EOLed for some time now, so it should be okay.