Expand references in twice-nested subschemas
If you have a schema that references a property that is defined
in another schema, and that property is a subschema, then the
subschema would not get expanded.

I added a loop to dereference subschemas, but this blew up, because the second
reference was local to the second schema, and did therefore not have a URI.
When the 'resolve_pointer' method did it's lookup, it would find no URI, and
attempt to resolve against the base schema '@schema' which is the wrong one.

By copying over the URI from the ref_schema, we get the lookup to happen
in the right place.