Handle expansion for nested external property
If you have a schema that has a reference to an object definition in a separate
schema, then there are three possibilities. The object properties in the second
schema could be:

1. actual properties
2. references to definitions within the same schema
3. references to definitions within a separate schema

When expanding references in the first schema, it would correctly handle both
cases 1 and 2, but not 3.

This adds a test for this use case, as well as the code to fix it.