remove filename gate, fixes #118
Remove the filename gate in validate-schema and update the test to use ruby's
"-c" option to test the syntax.

The filename gate in validate-schema causes problems when it is installed as a
binstub via RubyGems, since in this case the program will not execute. This
has been described in #118, and it is detailed in https://github.com/rbenv/rbenv/issues/1178